arreglos

+++ FrontEnd/src/views/Alumnos/AlumnosView.jsx
@@ -1,6 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { getAlumnos, addAlumno, updateAlumno, deleteAlumno } from '../../services/alumnosService';
 import './AlumnosView.css';
+import { FaPen,FaRegTrashAlt,FaPlusSquare  } from "react-icons/fa";
+import CreateAlumnoModal from './CreateAlumnoModal';
+
 
 function AlumnosView() {
     const [alumnos, setAlumnos] = useState([]);
@@ -11,6 +14,7 @@ function AlumnosView() {
     const [telefonoContacto, setTelefonoContacto] = useState('');
     const [correoElectronico, setCorreoElectronico] = useState('');
     const [editCi, setEditCi] = useState(null);
+    const [isOpen, setIsOpen] = useState(false);
 
     useEffect(() => {
         fetchAlumnos();
@@ -26,14 +30,14 @@ function AlumnosView() {
     };
 
     const handleAddOrUpdateAlumno = async (e) => {
-        e.preventDefault();
+        // e.preventDefault();
         const alumno = {
-            ci,
-            nombre,
-            apellido,
-            fecha_nacimiento: fechaNacimiento,
-            telefono_contacto: telefonoContacto,
-            correo_electronico: correoElectronico
+            ci: e.ci,
+            nombre: e.nombre,
+            apellido: e.apellido,
+            fecha_nacimiento: e.fechaNacimiento,
+            telefono_contacto: e.telefonoContacto,
+            correo_electronico: e.correoElectronico
         };
         try {
             if (editCi) {
@@ -49,6 +53,12 @@ function AlumnosView() {
     };
 
     const handleEdit = (alumno) => {
+        if(alumno==-1){
+            setIsOpen(true);
+            setEditCi(null);
+        }
+        else{
+        setIsOpen(true);
         setEditCi(alumno.ci);
         setCi(alumno.ci);
         setNombre(alumno.nombre);
@@ -56,14 +66,17 @@ function AlumnosView() {
         setFechaNacimiento(alumno.fecha_nacimiento);
         setTelefonoContacto(alumno.telefono_contacto);
         setCorreoElectronico(alumno.correo_electronico);
+    }
     };
 
     const handleDelete = async (ci) => {
-        try {
+        if (window.confirm("¿Estás seguro de que deseas eliminar este alumno?")) {
+            try {
             await deleteAlumno(ci);
             fetchAlumnos();
-        } catch (error) {
+            } catch (error) {
             console.error("Error al eliminar alumno:", error);
+            }
         }
     };
 
@@ -79,28 +92,53 @@ function AlumnosView() {
 
     return (
         <div className="alumnos-view">
-            <h1>Alumnos</h1>
+            <div class="title-container">
 
-            <form onSubmit={handleAddOrUpdateAlumno} className="alumno-form">
-                <input type="text" placeholder="CI" value={ci} onChange={(e) => setCi(e.target.value)} required disabled={!!editCi} />
-                <input type="text" placeholder="Nombre" value={nombre} onChange={(e) => setNombre(e.target.value)} required />
-                <input type="text" placeholder="Apellido" value={apellido} onChange={(e) => setApellido(e.target.value)} required />
-                <input type="date" placeholder="Fecha de Nacimiento" value={fechaNacimiento} onChange={(e) => setFechaNacimiento(e.target.value)} required />
-                <input type="text" placeholder="Teléfono de Contacto" value={telefonoContacto} onChange={(e) => setTelefonoContacto(e.target.value)} required />
-                <input type="email" placeholder="Correo Electrónico" value={correoElectronico} onChange={(e) => setCorreoElectronico(e.target.value)} required />
-                <button type="submit">{editCi ? "Actualizar" : "Agregar"} Alumno</button>
-                <button type="button" onClick={resetForm}>Cancelar</button>
-            </form>
+            <h1>Alumnos</h1> <button class="create-button" onClick={() => handleEdit(-1)}><FaPlusSquare  /> Crear nuevo</button>
+            </div>
 
-            <ul className="alumno-list">
-                {alumnos.map((alumno) => (
-                    <li key={alumno.ci} className="alumno-item">
-                        <span>{alumno.nombre} {alumno.apellido} - {alumno.ci}</span>
-                        <button onClick={() => handleEdit(alumno)}>Editar</button>
-                        <button onClick={() => handleDelete(alumno.ci)}>Eliminar</button>
-                    </li>
-                ))}
-            </ul>
+         <CreateAlumnoModal isOpen={isOpen}
+        onRequestClose={() => setIsOpen(false)}
+        initialData={{ ci, nombre, apellido, fechaNacimiento, telefonoContacto, correoElectronico }}
+        editCi={editCi}
+        handleAddOrUpdateAlumno={handleAddOrUpdateAlumno}
+        resetForm={resetForm}
+        
+         >
+
+        </CreateAlumnoModal>
+
+            <table className="alumno-table">
+                <thead>
+                    <tr>
+                        <th>CI</th>
+                        <th>Nombre</th>
+                        <th>Apellido</th>
+                        <th>Fecha de Nacimiento</th>
+                        <th>Teléfono de Contacto</th>
+                        <th>Correo Electrónico</th>
+                        <th>Acciones</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {alumnos.map((alumno) => (
+                        <tr key={alumno.ci}>
+                            <td>{alumno.ci}</td>
+                            <td>{alumno.nombre}</td>
+                            <td>{alumno.apellido}</td>
+                            <td>{new Date(alumno.fecha_nacimiento).toUTCString()}</td>
+                            <td>{alumno.telefono_contacto}</td>
+                            <td>{alumno.correo_electronico}</td>
+                            <td>
+                                <div class="Button-container">
+                                <button onClick={() => handleEdit(alumno)}><FaPen /> Editar</button>
+                                <button onClick={() => handleDelete(alumno.ci)}><FaRegTrashAlt /> Eliminar</button>
+                                </div>
+                            </td>
+                        </tr>
+                    ))}
+                </tbody>
+            </table>
         </div>
     );
 }
